+++ src/components/table/table_row_cell.tsx
@@ -18,6 +18,7 @@
  */
 
 import React, {
+  CSSProperties,
   Fragment,
   FunctionComponent,
   ReactElement,
@@ -35,6 +36,8 @@ import {
 } from '../../services';
 
 import { resolveWidthAsStyle } from './utils';
+
+import { useIsWithinBreakpoints } from '../../services/hooks/useIsWithinBreakpoints';
 
 interface EuiTableRowCellSharedPropsShape {
   /**
@@ -86,6 +89,10 @@ interface EuiTableRowCellMobileOptionsShape {
    * (typically cells are contained to 50%)
    */
   fullWidth?: boolean;
+  /**
+   * Applies the value to the width of the cell in mobile view (typically 50%)
+   */
+  width?: CSSProperties['width'];
 }
 
 interface EuiTableRowCellProps {
@@ -201,7 +208,12 @@ export const EuiTableRowCell: FunctionComponent<Props> = ({
     euiTableCellContent__hoverItem: showOnHover,
   });
 
-  const styleObj = resolveWidthAsStyle(style, width);
+  const widthValue =
+    useIsWithinBreakpoints(['xs', 's', 'm']) && mobileOptions.width
+      ? mobileOptions.width
+      : width;
+
+  const styleObj = resolveWidthAsStyle(style, widthValue);
 
   function modifyChildren(children: ReactNode) {
     let modifiedChildren = children;